# docs/game-out/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: '0.5'
  title: Game Out
  license:
    name: HAN
  description:
    Game Out is een web-applicatie waarmee buitenspelen en gamen gecombineerd wordt. 
host: server3.tezzt.nl:6031
basePath: /api
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /groups/{id}:
    get:
      summary: Get a single group by ID
      operationId: retrieve
      tags:
        - groups
      parameters:
        - name: id
          in: path
          description: ID of group to fetch
          required: true
          type: integer
          format: int64  
      responses:
        200:
          description: single group
          schema:
            $ref: '#/definitions/Group'
definitions:
...
  Group:
    required:
      - name
    properties:
      name:
        type: string
...